Refact SideMenu and Navbar to Dashboard component
